Fixed issues with prediction markets fetching from whitelisted accounts
