+++ app/components/PredictionMarkets/CreateMarketModal.jsx
@@ -3,7 +3,6 @@ import {Modal, Input, Form, Button} from "bitshares-ui-style-guide";
 import PropTypes from "prop-types";
 import Translate from "react-translate-component";
 import AssetSelect from "../Utility/AssetSelect";
-import AmountSelector from "../Utility/AmountSelectorStyleGuide";
 import counterpart from "counterpart";
 import AssetActions from "actions/AssetActions";
 import assetUtils from "common/asset_utils";
@@ -34,7 +33,7 @@ export default class CreateMarketModal extends Modal {
                     amount: 1
                 },
                 base: {
-                    asset_id: "1.3.0", //TODO
+                    asset_id: "1.3.0",
                     amount: 1
                 }
             },
@@ -57,41 +56,34 @@ export default class CreateMarketModal extends Modal {
         this.onSubmit = this.onSubmit.bind(this);
     }
 
-    _getPermissions() {
+    _getPermissionsAndFlags() {
         let flagBooleans = assetUtils.getFlagBooleans(0, IS_BITASSET);
         let permissionBooleans = assetUtils.getFlagBooleans("all", IS_BITASSET);
 
         flagBooleans["charge_market_fee"] = true;
-
+        let flags = assetUtils.getFlags(flagBooleans, IS_BITASSET);
         return {
-            flagBooleans,
-            permissionBooleans
+            flags,
+            permissions: assetUtils.getPermissions(
+                permissionBooleans,
+                IS_BITASSET
+            )
         };
     }
 
-    _createAsset(e) {
-        if (e) {
-            e.preventDefault();
-        }
-        const {flagBooleans, permissionBooleans} = this._getPermissions();
-
+    _createAsset() {
         let {marketOptions, core_exchange_rate, bitasset_opts} = this.state;
 
-        const accountId = ChainStore.getAccount(this.props.currentAccount).get(
-            "id"
-        );
-
-        let flags = assetUtils.getFlags(flagBooleans, IS_BITASSET);
-        let permissions = assetUtils.getPermissions(
-            permissionBooleans,
-            IS_BITASSET
-        );
-
+        const {permissions, flags} = this._getPermissionsAndFlags();
         const description = JSON.stringify(
             this.state.marketOptions.description
         );
+
         this.setState({inProgress: true});
-        const creationPromise = AssetActions.createAsset(
+        const accountId = ChainStore.getAccount(this.props.currentAccount).get(
+            "id"
+        );
+        AssetActions.createAsset(
             accountId,
             marketOptions,
             flags,
@@ -101,8 +93,7 @@ export default class CreateMarketModal extends Modal {
             true,
             bitasset_opts,
             description
-        );
-        creationPromise
+        )
             .then(result => {
                 this.setState({inProgress: false});
                 console.log(
@@ -121,21 +112,24 @@ export default class CreateMarketModal extends Modal {
     }
 
     handleChange(event) {
-        let newMarket = this.state.marketOptions;
+        let marketOptions = this.state.marketOptions;
         switch (event.target.name) {
             case "symbol":
-                newMarket[event.target.name] = event.target.value.toUpperCase();
+                marketOptions[
+                    event.target.name
+                ] = event.target.value.toUpperCase();
                 break;
             case "main":
             case "condition":
             case "expiry":
-                newMarket.description[event.target.name] = event.target.value;
+                marketOptions.description[event.target.name] =
+                    event.target.value;
                 break;
             default:
-                newMarket[event.target.name] = event.target.value;
+                marketOptions[event.target.name] = event.target.value;
                 break;
         }
-        this.setState({marketOptions: newMarket});
+        this.setState({marketOptions});
     }
 
     handleAssetChange(asset) {
@@ -200,7 +194,10 @@ export default class CreateMarketModal extends Modal {
 
     onSubmit(e) {
         if (this._isFormValid()) {
-            this._createAsset.call(this, e);
+            if (e) {
+                e.preventDefault();
+            }
+            this._createAsset().call(this);
         } else {
             this.setState({showWarning: true});
         }
